[prim/arbiter] Keep current request until acked
If downstream isn't ready (arb_ready = 0), then if higher priority
request comes later but before ready, it outtakes previous request,
which results in the arb_data change.

It doesn't harm as the data supposes to be used when valid & ready. It,
however, looks bad on the crossbar. inside `tlul_socket_m1` it could
happen that the downstream request write data or address could be
changed without notice, and it violates TL-UL spec.

This is related to #152