[prim/packer] Enhance the behavior based on FPV result
Problem 1:

    `prim_packer` did not assert `valid_o` even it has enough data to
    send out on some corner case when `ready_i` is de-asserted.

Lost the previous commit message. So let it be short. `prim_packer` is
a primitive module to pack incoming data with mask into `OutW` bit-width
signal and send out to the output port.

While preparing FPV environment, a new assertion was added:

    // If input mask is greater than output width, valid should be
    // asserted
    `ASSERT(ValidOAssertedForInputGTEOutW_A,
            valid_i && ($countones(mask_i) >= OutW) |-> valid_o,
            clk_i, !rst_ni)

FPV catches a corner case based on this assertion. When the output port
lowered `ready_i`, the module stores incoming request into the internal
registers. Then it increases `pos` pointer value by `InW`. At the next
cycle, it is expected to keep asserting the `valid_o` as it has data to
send out. As `pos_next` is wrapped around in next cycle if there is more
incoming data, the `valid_next` is lowered.

Resolution:

    Now `prim_packer` considers `pos` value to assert `valid_o` prior to
    compare `pos_next`.

Problem 2:

    `ready_o` falsely asserted as `pos_next` wrapped around.

This issue is similar with the Problem 1. Problem 1 didn't cause any
functionally incorrect behavior. This problem causes an issue. As it let
the design latch the data even doesn't have sufficient storage inside.

Resolution:

    `ready_next` logic doesn't look `pos_next` but only looks at `pos`.
    If `pos` is less than `OutW`, it has enough internal storage to
    store incoming packet. Also, remind that if output port acked the
    outgoing data, it can always store incoming data.

This is related to #19