[rv_plic/fpv] This aligns the naming scheme and hierarchy of all FPV TBs
This change aligns the hierarchy and naming scheme of all FPV TBs.

In particular:

- FPV testbenches are named `<design>_fpv.sv`
- FPV assert files are named `<design>_assert_fpv.sv`
- FPV bind files are named `<design>_bind_fpv.sv`

This renaming of FPV-related files helps to distinguish them from the
simulation-based counterparts. This change intends to avoid name collisions and
accidental replacement of designs during compilation.

Further, the FPV testbenches of designs that are dependent on the top-level
configuration have been adapted to include the top-level parameterization.

The `rv_plic` design has further been modified such that it uses the constants
generated by the reggen tool for consistency. Also, an `rv_plic_fpv` testbench
has been added which instantiates different configurations of the `rv_plic`
design.