[prim] Make ROM_INIT_FILE a parameter
Instead of having a global define ROM_INIT_FILE, which contains a path
to a VMEM file with the boot ROM contents, use a parameter passed
through the hierarchy. This enables multiple instantiations of the ROM,
and avoids naming clashes between the ROM_INIT_FILE/SRAM_INIT_FILE
define, which can be used only once in the whole toplevel design.

This requires changes in various places to properly pass through the
parameter.

Instead of updating the prim_xilinx_rom variant, remove it. This
primitive is identical to the generic one, with the addition of
generating an empty ROM if no ROM initialization file is present during
the synthesis run. We capture this problem now in a more user-friendly
way with a TCL script after synthesis; removing this primitive here
avoids updating it first to the new way of passing in the init file.

The Arty S7 changes were done blindly, I don't have that board. I also
note that the board was very likely broken before, as the ROM init file
name was not called SRAM_INIT_FILE for a long time.

Fixes #2225

Signed-off-by: Philipp Wagner <phw@lowrisc.org>