[prim] Improve extraction of parameter port list
The `_parse_parameter_port_list()` function in primgen extracts the
names of parameters from a parameter port list. The extraction failed
for parameters without an initial value, or with a `-` in the data type
definition (as in `Width-1`). Fix both cases by updating the regex.

This commit also adds a small inline test, to be run with `pytest`, to
make it easier to improve this code in the future.

Fixes #2679

Signed-off-by: Philipp Wagner <phw@lowrisc.org>