[Keccak] Keccak_f implementation
This commit implements the full round of Keccak computation. It supports
all configurations of Width from 25 to 1600 bits. The number of rounds
is decided based on the given input width. For example, 800 bit Width
runs 22 rounds.

This implementation has the keccak states registers inside the module.
So, it has the message feed interface to support SHA3/SHAKE. The input
interface width is also configurable. But `Width`/`DInWidth` shall be
integer. For instance, 50b `Width` case, the `DInWidth` can be 25b but
not 27b.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>