[pinmux/padring] Wire up the pad attribute WARL behavior modules
This connects the WARL modules for the pad attribute CSRs in the pinmux.
Note that the pad type information has to be passed down as a parameter
from the chiplevel where the padring templating is implemented.

Note that the pad / peripheral signal correspondence is different for
muxed pads than for dedicated ones.

Muxed peripheral signals first go into the pinmux array, which in turn
is connected to an array of muxed pads.  The muxed pad attribute CSRs
directly correspond to that array of pads, so figuring out the correct
pad type for CSR WARL configuraration purposes is straightforward.

For dedicated peripheral signals, the story is a bit different. In
general each directly connected peripheral signal can be directly
associated with the pad, in which case pad type inference is easy.

However, we do have a couple of manually connected dedicated peripheral
signals in the system (mostly USB), and the pad type for CSR WARL
configuration cannot be automatically inferred in that case.  Hence, we
provide that information via an additional flag in the pinmux Hjson
configuration.

Signed-off-by: Michael Schaffner <msf@google.com>