[otbn/otp_ctrl] Replicate dmem scrambling keystream
This replicates the keystream instead of instantiating multiple PRINCE
primitives, as discussed in detail in #7054.
This effectively means that the OTBN nonce is now only 64bit wide, and
this patch changes the widths throughout the otp_ctrl and otbn designs.

This also aligns the OTP control DV environment to only expect 10
instead of 16 EDN fetch requests.

Signed-off-by: Michael Schaffner <msf@opentitan.org>
Signed-off-by: Cindy Chen <chencindy@opentitan.org>